Adding Validation
Added FluentValidation Package to our project.
Created Validators for Dtos which we using with our commands (create, update etc.)
Created interfaces inside Dto folder to write common properties to inside of the folder so when we validate that interface we wont be need to rewrite
same rules for same properties over and over again.